refactor: 움직임 전략 패키지 분리
