+++ src/main/java/janggi/domain/piece/line/CannonMoveStrategy.java
@@ -1,9 +1,10 @@
-package janggi.domain.piece;
+package janggi.domain.piece.line;
 
 import static janggi.domain.piece.PieceType.CANNON;
 
 import janggi.domain.board.BoardSnapshot;
 import janggi.domain.dynasty.Dynasty;
+import janggi.domain.piece.MoveStrategy;
 import janggi.domain.position.Direction;
 import janggi.domain.position.Position;
 import java.util.ArrayList;
